rewrote the error checking for if the file exist and writing csv files to a dictionary so that they were classes instead of just procedural.
