+++ .gitignore
@@ -4,3 +4,4 @@ __pycache__
 *.pyc
 *.csv
 .DS_Store
+network_images/
