+++ write_path.py
@@ -1,5 +1,4 @@
 import os
-import logging
 
 
 def writePath(write_location, file_name):
